% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/center_around.R
\name{center_around}
\alias{center_around}
\title{Center around a Coordinate}
\usage{
center_around(df, x_cent, y_cent, coord1 = x, coord2 = y)
}
\arguments{
\item{coord1}{optional, only when different from x}

\item{coord2}{optional, only when different from y}
}
\description{
Center around a Coordinate
}
