% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/addins.R
\name{insertsummarize}
\alias{insertsummarize}
\title{Insert dplyr::select(.}
\usage{
insertsummarize()
}
\description{
Call this function as an addin to insert \code{ dplyr::summarize( } at the cursor position.
}
