% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_circle.R
\name{make_circle}
\alias{make_circle}
\title{Title}
\usage{
make_circle(r = 1, start = 0, l = 100)
}
\arguments{
\item{r}{radius}

\item{l}{number of points}
}
\value{
a tibble with variable x and y with circle coordinates
}
\description{
Title
}
\examples{
make_circle() \%>\% quick
}
