% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/make_circle.R
\name{make_circles}
\alias{make_circles}
\title{Data Frame of Circles}
\usage{
make_circles(r_vec, start = 0, l = 100)
}
\arguments{
\item{l}{}
}
\description{
Data Frame of Circles
}
\examples{
make_circles(1:100)
}
